requirements.txt
aiohappyeyeballs==2.3.5
aiohttp==3.10.2
aiosignal==1.3.1
anyio==4.3.0
async-timeout==4.0.3
nest-asyncio
attrs==23.2.0
azure-core==1.31.0
azure-identity==1.19.0
certifi==2024.7.4
cffi==1.16.0
charset-normalizer==3.3.2
cryptography==42.0.5
deprecated==1.2.14
exceptiongroup==1.2.0
Flask==3.1.0
frozenlist==1.4.1
google-generativeai
google-genai==0.3.0
grpcio==1.60.1
h11==0.14.0
h2==4.1.0
hpack==4.0.0
httpcore==1.0.4
httpx[http2]==0.27.0
hyperframe==6.0.1
idna==3.7
importlib-metadata==6.11.0
llama-index-llms-gemini
llama-index-embeddings-gemini
microsoft-kiota-abstractions==1.3.1
microsoft-kiota-authentication-azure==1.0.0
microsoft-kiota-http==1.3.1
microsoft-kiota-serialization-form==0.1.0
microsoft-kiota-serialization-json==1.3.0
microsoft-kiota-serialization-multipart==0.1.0
microsoft-kiota-serialization-text==1.0.0
msal==1.31.0
msal-extensions==1.2.0
msgraph-core==1.0.0
msgraph-sdk==1.14.0
multidict==6.0.5
opentelemetry-api==1.23.0
opentelemetry-sdk==1.23.0
opentelemetry-semantic-conventions==0.44b0
pendulum==3.0.0
portalocker==2.8.2
pycparser==2.21
pyjwt[crypto]==2.8.0
python-dateutil==2.9.0.post0
requests==2.32.0
six==1.16.0
sniffio==1.3.1
std-uritemplate==0.0.54
streamlit==1.41.1
time-machine==2.14.0
typing-extensions==4.10.0
tzdata==2024.1
urllib3==2.2.2
wrapt==1.16.0
yarl==1.9.4
zipp==3.19.1